rethinking export button functions
