+++ my-school/src/Components/MainLayout.js
@@ -27,7 +27,7 @@ const MainLayout = ({ page, getFamily, getUserByID }) => {
 }
 
 const mapStateToProps = (state) => {
-    console.log("state in main layout", state)
+    // console.log("state in main layout", state)
     return {
         user: state.usersReducer.user,
         family: state.usersReducer.family
