+++ my-school/src/Components/studentRegister.js
@@ -1,6 +1,6 @@
 import React, { useRef } from "react";
 import { useForm } from "react-hook-form";
-import { axiosWithAuth } from "../utils/axiosWithAuth.js";
+// import { axiosWithAuth } from "../utils/axiosWithAuth.js";
 import { connect } from "react-redux";
 import { getFamily } from "../Redux/actions/actions-users";
 import {
